Move duplicated function to common parent class Fix unit tests
