+++ resources/lib/pyscraper/descriptionparserxml.py
@@ -76,74 +76,6 @@ class DescriptionParserXml(DescriptionParser):
 			result = self.parseElement(node)
 			result = self.replaceResultTokens(result)
 			yield result
-	
-	
-	#TODO: make a base class and make this a base method
-	def replaceResultTokens(self, resultAsDict):
-				
-		for key in resultAsDict.keys():
-			grammarElement = self.grammarNode.find(key)
-			if(grammarElement != None):
-				appendResultTo = grammarElement.attrib.get('appendResultTo')
-				appendResultWith = grammarElement.attrib.get('appendResultWith')
-				replaceKeyString = grammarElement.attrib.get('replaceInResultKey')
-				replaceValueString = grammarElement.attrib.get('replaceInResultValue')
-				dateFormat = grammarElement.attrib.get('dateFormat')
-				del grammarElement
-														
-				#TODO: avoid multiple loops
-				if(appendResultTo != None or appendResultWith != None or dateFormat != None):									
-					itemList = resultAsDict[key]
-					for i in range(0, len(itemList)):
-						try:
-							item = itemList[i]
-							newValue = item
-							del item
-							if(appendResultTo != None):								
-								newValue = appendResultTo +newValue
-							if(appendResultWith != None):
-								newValue = newValue + appendResultWith
-							if(dateFormat != None):
-								if(dateFormat == 'epoch'):
-									try:
-										newValue = time.gmtime(int(newValue))
-									except:
-										print 'error converting timestamp: ' +str(newValue) 
-								else:
-									newValue = time.strptime(newValue, dateFormat)
-							itemList[i] = newValue
-						except Exception, (exc):
-							print "Error while handling appendResultTo: " +str(exc)
-							
-					resultAsDict[key] = itemList
-					del itemList
-					
-				if(replaceKeyString != None and replaceValueString != None):												
-					replaceKeys = replaceKeyString.split(',')
-					replaceValues = replaceValueString.split(',')
-					
-					if(len(replaceKeys) != len(replaceValues)):
-						print "Configuration error: replaceKeys must be the same number as replaceValues"
-					
-					itemList = resultAsDict[key]
-					for i in range(0, len(itemList)):
-						try:							
-							item = itemList[i]
-							
-							for j in range(len(replaceKeys)):
-								replaceKey = replaceKeys[j]
-								replaceValue = replaceValues[j]
-															
-								newValue = item.replace(replaceKey, replaceValue)
-								del item							
-								itemList[i] = newValue
-						except:
-							print "Error while handling appendResultTo"
-							
-					resultAsDict[key] = itemList
-					del itemList
-		
-		return resultAsDict			
 
 			
 	def parseElement(self, sourceTree):
